refactor(compare): remove pure html report
