--- projects/compare/src/compare/templates/base.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>{% block title %}Database Comparison Report{% endblock %}</title>
-    {% include 'styles.html' %}
-</head>
-<body>
-    <div class="ct">
-        {% block content %}{% endblock %}
-    </div>
-    {% include 'scripts.html' %}
-</body>
-</html>
