--- projects/compare/src/compare/templates/comparison_report.html
@@ -1,57 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<div class="h">
-    <h1>Database Comparison Report</h1>
-    <div class="md">
-        <div><strong>Source:</strong> {{ comparison.source }}</div>
-        <div><strong>Target:</strong> {{ comparison.target }}</div>
-    </div>
-</div>
-
-<div class="s">
-    {% set total_tables = comparison.changes | length %}
-    {% set tables_with_changes = [] %}
-    {% for table in comparison.changes %}
-        {% if (table.schema | length) > 0 or (table.data | length) > 0 %}
-            {% set _ = tables_with_changes.append(table) %}
-        {% endif %}
-    {% endfor %}
-    
-    {% set ns = namespace(total_schema_changes=0, total_data_changes=0) %}
-    {% for table in comparison.changes %}
-        {% set ns.total_schema_changes = ns.total_schema_changes + (table.schema | length) %}
-        {% set ns.total_data_changes = ns.total_data_changes + (table.data | length) %}
-    {% endfor %}
-    
-    <div class="sc">
-        <h3>Tables Compared</h3>
-        <div class="n">{{ total_tables }}</div>
-    </div>
-    <div class="sc">
-        <h3>Tables with Changes</h3>
-        <div class="n">{{ tables_with_changes | length }}</div>
-    </div>
-    <div class="sc">
-        <h3>Schema Changes</h3>
-        <div class="n">{{ ns.total_schema_changes }}</div>
-    </div>
-    <div class="sc">
-        <h3>Data Changes</h3>
-        <div class="n">{{ ns.total_data_changes }}</div>
-    </div>
-</div>
-
-<div class="se">
-    <h2>Table Comparisons</h2>
-    <div class="tcs">
-        {% for table in comparison.changes %}
-            {% if (table.schema | length) > 0 or (table.data | length) > 0 %}
-                {% include 'table_comparison.html' %}
-            {% endif %}
-        {% else %}
-            <div class="nc">No changes detected</div>
-        {% endfor %}
-    </div>
-</div>
-{% endblock %}
