+++ projects/compare/src/compare/templates/json_report.html
@@ -1,0 +1,51 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Database Comparison Report</title>
+    {% include 'json_styles.html' %}
+</head>
+<body>
+    <div class="ct">
+        <div class="h">
+            <h1>Database Comparison Report</h1>
+            <div class="md">
+                <div><strong>Source:</strong> {{ comparison.source }}</div>
+                <div><strong>Target:</strong> {{ comparison.target }}</div>
+            </div>
+        </div>
+
+        <div class="s">
+            <div class="sc">
+                <h3>Tables Compared</h3>
+                <div class="n" id="total-tables">-</div>
+            </div>
+            <div class="sc">
+                <h3>Tables with Changes</h3>
+                <div class="n" id="tables-with-changes">-</div>
+            </div>
+            <div class="sc">
+                <h3>Schema Changes</h3>
+                <div class="n" id="schema-changes">-</div>
+            </div>
+            <div class="sc">
+                <h3>Data Changes</h3>
+                <div class="n" id="data-changes">-</div>
+            </div>
+        </div>
+
+        <div class="se">
+            <h2>Table Comparisons</h2>
+            <div class="tcs" id="tables-container">
+                <div class="loading">Loading tables...</div>
+            </div>
+        </div>
+    </div>
+
+    <script>
+        window.comparisonData = {{ comparison_json | safe }};
+        {% include 'json_scripts.js' %}
+    </script>
+</body>
+</html>
