+++ projects/compare/src/compare/templates/json_styles.html
@@ -62,14 +62,15 @@
         border-radius: 6px;
         margin-bottom: 20px;
     }
-    
     .tct {
         overflow-x: auto;
         margin: 15px 0;
         border: 1px solid #dee2e6;
         border-radius: 6px;
+        position: relative;
+        max-height: 400px;
+        overflow-y: auto;
     }
-    
     .dt, .st {
         width: 100%;
         min-width: 600px;
@@ -77,7 +78,6 @@
         margin: 0;
         font-size: 14px;
     }
-    
     .dt th, .st th {
         background: #f8f9fa;
         padding: 12px 16px;
@@ -86,8 +86,10 @@
         font-weight: 600;
         color: #495057;
         font-size: 13px;
+        position: sticky;
+        top: 0;
+        z-index: 10;
     }
-    
     .dt td, .st td {
         padding: 10px 16px;
         border: 1px solid #dee2e6;
@@ -100,39 +102,17 @@
         position: relative;
         cursor: help;
     }
-    
     .dt tr:nth-child(even), .st tr:nth-child(even) {
         background: #f8f9fa;
     }
-    
     .dt tr:hover, .st tr:hover {
         background: #e3f2fd;
     }
-    
-    .dt td:hover::after, .st td:hover::after {
-        content: attr(title);
-        position: absolute;
-        top: -5px;
-        left: 0;
-        background: #333;
-        color: white;
-        padding: 4px 8px;
-        border-radius: 4px;
-        font-size: 12px;
-        white-space: nowrap;
-        z-index: 1000;
-        box-shadow: 0 2px 8px rgba(0,0,0,0.3);
-        max-width: 300px;
-        word-wrap: break-word;
-        white-space: normal;
-    }
-    
     .null {
         color: #9ca3af;
         font-style: italic;
         font-weight: 300;
     }
-    
     .mc {
         background: #fff3cd !important;
         font-family: monospace;
@@ -143,7 +123,6 @@
         overflow: visible !important;
         text-overflow: unset !important;
     }
-    
     .ov {
         color: #dc3545;
         background: #f8d7da;
@@ -157,7 +136,6 @@
         text-overflow: ellipsis;
         vertical-align: middle;
     }
-    
     .nv {
         color: #155724;
         background: #d4edda;
@@ -171,7 +149,6 @@
         text-overflow: ellipsis;
         vertical-align: middle;
     }
-    
     .ar {
         color: #6c757d;
         font-weight: bold;
@@ -180,7 +157,6 @@
         display: inline-block;
         vertical-align: middle;
     }
-    
     .cb {
         padding: 4px 8px;
         border-radius: 4px;
@@ -191,7 +167,6 @@
     .cb.a { background: #d4edda; color: #155724; }
     .cb.r { background: #f8d7da; color: #721c24; }
     .cb.m { background: #fff3cd; color: #856404; }
-    
     .ca { background: rgba(212, 237, 218, 0.3) !important; }
     .cr { background: rgba(248, 215, 218, 0.3) !important; }
     .cm { background: rgba(255, 243, 205, 0.3) !important; }
@@ -223,23 +198,26 @@
         text-align: center;
         padding: 40px;
     }
-    .lg {
-        display: flex;
-        gap: 15px;
-        margin-bottom: 15px;
-        padding: 10px;
-        background: #f8f9fa;
-        border-radius: 4px;
+    .loading {
+        color: #666;
+        font-style: italic;
+        text-align: center;
+        padding: 20px;
+    }
+    .virtual-table {
+        min-height: 200px;
+    }
+    .virtual-row {
+        display: table-row;
+    }
+    .search-box {
+        margin: 10px 0;
+        padding: 8px 12px;
         border: 1px solid #dee2e6;
+        border-radius: 4px;
+        font-size: 14px;
+        width: 300px;
     }
-    .li {
-        display: flex;
-        align-items: center;
-        gap: 6px;
-        font-size: 13px;
-        font-weight: 500;
-    }
-    .tcs {}
     h4 .cb {
         margin-left: 8px;
         margin-right: 4px;
