--- projects/compare/src/compare/templates/scripts.html
@@ -1,13 +1,0 @@
-<script>
-    function toggleCollapsible(element) {
-        const content = element.nextElementSibling;
-        element.classList.toggle('expanded');
-        content.classList.toggle('show');
-    }
-
-    // Start with all sections collapsed
-    document.addEventListener('DOMContentLoaded', function() {
-        // All sections start collapsed by default - no auto-expansion
-        console.log('Database comparison report loaded with all sections collapsed');
-    });
-</script>
