+++ src/dpm_toolkit/cli.py
@@ -298,9 +298,9 @@ def create_parser() -> ArgumentParser:
     )
     compare_parser.add_argument(
         "--format",
-        choices=["json", "html", "html-js"],
+        choices=["json", "html"],
         default="json",
-        help="Output format: json (default), html (server-side), or html-js (client-side)",
+        help="Output format: json (default) or html",
     )
     compare_parser.add_argument(
         "--output",
@@ -453,7 +453,7 @@ def handle_schema_command(args: Namespace) -> None:
 def handle_compare_command(args: Namespace) -> None:
     """Handle the 'compare' subcommand."""
     try:
-        from compare import HtmlReportGenerator, JsonHtmlReportGenerator, compare_databases, comparison_to_json
+        from compare import JsonHtmlReportGenerator, compare_databases, comparison_to_json
     except ImportError as e:
         print(f"Compare functionality not available: {e}")
         return
@@ -481,16 +481,10 @@ def handle_compare_command(args: Namespace) -> None:
         json_output = comparison_to_json(result)
 
         # Handle output based on format
-        if args.format == "html-js":
+        if args.format == "html":
             # Generate JavaScript-based HTML report
             output_path = args.output or Path("comparison_report.html")
             generator = JsonHtmlReportGenerator()
-            generator.generate_report(result, output_path)
-            print(f"JavaScript HTML report saved to: {output_path}")
-        elif args.format == "html":
-            # Generate traditional server-side HTML report
-            output_path = args.output or Path("comparison_report.html")
-            generator = HtmlReportGenerator()
             generator.generate_report(result, output_path)
             print(f"HTML report saved to: {output_path}")
         elif hasattr(args, "output") and args.output:
